chore: rename /methods to /app/methods
